+++ src/view/film-card/film-card.js
